Use cached ism addresses for non core apps
