+++ typescript/infra/config/environments/testnet3/core.ts
@@ -1,9 +1,4 @@
-import {
-  AggregationIsmConfig,
-  ChainMap,
-  CoreConfig,
-  objMap,
-} from '@hyperlane-xyz/sdk';
+import { ChainMap, CoreConfig, objMap } from '@hyperlane-xyz/sdk';
 
 import { aggregationIsm } from '../../aggregationIsm';
 import { Contexts } from '../../contexts';
@@ -11,11 +6,7 @@ import { Contexts } from '../../contexts';
 import { owners } from './owners';
 
 export const core: ChainMap<CoreConfig> = objMap(owners, (local, owner) => {
-  const defaultIsm: AggregationIsmConfig = aggregationIsm(
-    'testnet3',
-    local,
-    Contexts.Hyperlane,
-  );
+  const defaultIsm = aggregationIsm('testnet3', local, Contexts.Hyperlane);
   return {
     owner,
     defaultIsm,
